Split the proxy behavior from the encapsulation in a servlet
This is in the purpose of using the proxy in an independent way.
