+++ src/main/java/org/mitre/dsmiley/httpproxy/Logger.java
@@ -1,0 +1,6 @@
+package org.mitre.dsmiley.httpproxy;
+
+public interface Logger {
+  void logMessage(String s, Exception e);
+  void logMessage(String s);
+}
