+++ src/main/java/org/mitre/dsmiley/httpproxy/QueryStringRewriter.java
@@ -1,0 +1,7 @@
+package org.mitre.dsmiley.httpproxy;
+
+import javax.servlet.http.HttpServletRequest;
+
+public interface QueryStringRewriter {
+  String rewriteQueryStringFromRequest(HttpServletRequest servletRequest, String queryString);
+}
